requirements.txt
streamlit==1.26.0
opencv-python-headless==4.7.0.72
Pillow==9.5.0
scikit-image==0.20.0
pandas==2.0.3
numpy==1.24.3